requirements.txt
Flask
gunicorn
openai==0.28.1
python-dateutil
python-dotenv
reportlab
qrcode
requests
